[FIX] Separando classes em arquivos diferentes
