+++ l10n_br_resource/models/__init__.py
@@ -3,3 +3,4 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 from . import resource_calendar
+from . import resource_calendar_leaves
