+++ l10n_br_resource/models/resource_calendar.py
@@ -235,33 +235,3 @@ class ResourceCalendar(models.Model):
             return 30
         else:
             return quantidade_dias
-
-
-class ResourceCalendarLeave(models.Model):
-
-    _inherit = 'resource.calendar.leaves'
-
-    country_id = fields.Many2one(
-        'res.country', string=u'País',
-        related='calendar_id.country_id',
-    )
-    state_id = fields.Many2one(
-        'res.country.state', u'Estado',
-        related='calendar_id.state_id',
-        domain="[('country_id','=',country_id)]",
-        readonly=True
-    )
-    l10n_br_city_id = fields.Many2one(
-        'l10n_br_base.city', u'Municipio',
-        related='calendar_id.l10n_br_city_id',
-        domain="[('state_id','=',state_id)]",
-        readonly=True
-    )
-    leave_type = fields.Selection(
-        string=u'Tipo',
-        selection=[item for item in TIPO_FERIADO.iteritems()],
-    )
-    abrangencia = fields.Selection(
-        string=u'Abrangencia',
-        selection=[item for item in ABRANGENCIA_FERIADO.iteritems()],
-    )
