--- l10n_br_resource/models/resource_resource.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright 2016 KMEE - Luis Felipe Miléo <mileo@kmee.com.br>
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
-from openerp import models
-
-
-class ResourceResource(models.Model):
-
-    _inherit = 'resource.resource'
